sorting of post and validations added

+++ routes/comments.js
@@ -19,25 +19,20 @@ router.post('/',verifyToken,async(req,res)=>{
         return res.status(400).send({message:error['details'][0]['message']})
     }
 
-    // Validation to check if owner exists
-    const userExists = await User.findOne({email:req.body.commentOwner})
-    if(!userExists){
-        return res.status(400).send({message:'Comment owner doesnot exist, provide a valid user'})
-    }
-
     const postExists = await Post.findById(req.body.postId)
     if(!postExists){
         return res.status(400).send({message:'Post doesnot exist, provide a valid post'})
     }
 
-   const postOwnerEqualsPostOwner = postExists['postOwner'] === req.body.commentOwner
-    if(postOwnerEqualsPostOwner){
+    const user = await User.findById(req.user._id)
+    const postOwnerEqualsUser = postExists['postOwner'] === user['email']
+    if(postOwnerEqualsUser){
         return res.status(400).send({message:'Post Owner canot comment on his own post'})
     }
 
     const commentData = new Comment({
         postId:req.body.postId,
-        commentOwner:req.body.commentOwner,
+        commentOwner:user['email'],
         commentText:req.body.commentText
     })
     try{
@@ -49,27 +44,29 @@ router.post('/',verifyToken,async(req,res)=>{
 })
 
 // Read all
-router.get('/:commentOwner', verifyToken, async(req,res) =>{
+router.get('/', verifyToken, async(req,res) =>{
+    const user = await User.findById(req.user._id)
     try{
-        const comments = await Comment.find({commentOwner:req.params.commentOwner})
+        const comments = await Comment.find({commentOwner:user['email']})
         res.send(comments)
     }catch(err){
         res.status(400).send({message:err})
     }
 })
 
-// Read by id
-router.get('/:commentId',verifyToken, async(req,res) =>{
-    try{
-        const getCommentById = await Comment.findById(req.params.commentId)
-        res.send(getCommentById)
-    }catch(err){
-        res.send({message:err})
-    }
-})
 
 // Update Comment
 router.patch('/:commentId',verifyToken, async(req,res) =>{
+
+    const user = await User.findById(req.user._id)
+    const comment = await Comment.findById(req.params.commentId)
+    if(!comment){
+        return res.status(400).send({message:'Comment doesnot exist,please provide a valid Comment'})
+    }
+    const postOwnerNotEqualsUser = comment['commentOwner'] !== user['email']
+    if(postOwnerNotEqualsUser){
+        return res.status(400).send({message:'Comment Owner can only update comment'})
+    }
     try{
         const updateCommentById = await Comment.updateOne(
             {_id:req.params.commentId},
@@ -87,6 +84,15 @@ router.patch('/:commentId',verifyToken, async(req,res) =>{
 
 // delete Comment
 router.delete('/:commentId',verifyToken,async(req,res)=>{
+    const user = await User.findById(req.user._id)
+    const comment = await Comment.findById(req.params.commentId)
+    if(!comment){
+        return res.status(400).send({message:'Comment doesnot exist,please provide a valid Comment'})
+    }
+    const postOwnerNotEqualsUser = comment['commentOwner'] !== user['email']
+    if(postOwnerNotEqualsUser){
+        return res.status(400).send({message:'Comment Owner can only delete comment'})
+    }
     try{
         const deleteCommentById = await Comment.deleteOne({_id:req.params.commentId})
         res.send(deleteCommentById)
